Massive changes to autonomous on 1/12/17
AutoBluePos1.java now works for hitting both beacons and simulating the shots at the center vortex.
AutoRedPos1.java is now modified to be a mirror of AutoBluePos1.java, but more tuning is needed on the drive distances.

This is expected to be the last change until after 1/18/17 due to the builders needing time to work on the robot.